+++ mex/mex.py
@@ -8,14 +8,14 @@ from reflex.utils.console import info as log_info
 
 from mex.common.logging import logger
 from mex.editor.api.main import check_system_status, get_prometheus_metrics
-from mex.editor.aux_search.main import index as aux_search_index
-from mex.editor.aux_search.state import AuxState
 from mex.editor.consent.main import index as consent_index
 from mex.editor.consent.state import ConsentState
 from mex.editor.create.main import index as create_index
 from mex.editor.create.state import CreateState
 from mex.editor.edit.main import index as edit_index
 from mex.editor.edit.state import EditState
+from mex.editor.ingest.main import index as ingest_index
+from mex.editor.ingest.state import IngestState
 from mex.editor.login_ldap.main import index as login_ldap_index
 from mex.editor.login_mex.main import index as login_mex_index
 from mex.editor.rules.state import RuleState
@@ -28,19 +28,6 @@ app = rx.App(
     html_lang="en",
     theme=themes.theme(accent_color="blue", has_background=False),
     style={">a": {"opacity": "0"}},
-)
-app.add_page(
-    edit_index,
-    route="/item/[identifier]",
-    title="MEx Editor | Edit",
-    on_load=[
-        State.check_login,
-        State.load_nav,
-        RuleState.refresh,
-        EditState.load_item_title,
-        EditState.show_submit_success_toast_on_redirect,
-        RuleState.resolve_identifiers,
-    ],
 )
 app.add_page(
     search_index,
@@ -56,17 +43,6 @@ app.add_page(
     ],
 )
 app.add_page(
-    aux_search_index,
-    route="/aux-search",
-    title="MEx Editor | Aux Search",
-    on_load=[
-        State.check_login,
-        State.load_nav,
-        AuxState.refresh,
-        AuxState.resolve_identifiers,
-    ],
-)
-app.add_page(
     create_index,
     route="/create",
     title="MEx Editor | Create",
@@ -76,6 +52,30 @@ app.add_page(
         CreateState.reset_stem_type,
         RuleState.refresh,
         RuleState.resolve_identifiers,
+    ],
+)
+app.add_page(
+    edit_index,
+    route="/item/[identifier]",
+    title="MEx Editor | Edit",
+    on_load=[
+        State.check_login,
+        State.load_nav,
+        RuleState.refresh,
+        EditState.load_item_title,
+        EditState.show_submit_success_toast_on_redirect,
+        RuleState.resolve_identifiers,
+    ],
+)
+app.add_page(
+    ingest_index,
+    route="/ingest",
+    title="MEx Editor | Ingest",
+    on_load=[
+        State.check_login,
+        State.load_nav,
+        IngestState.refresh,
+        IngestState.resolve_identifiers,
     ],
 )
 app.add_page(